Use strict bash mode in publish.sh
This will ensure that if pandoc is missing or fails
that the upload to PyPI is aborted.

Otherwise it would proceed but use README.md rather
than README.rst for the long description, which would
not render nicely on PyPI which expects the text to
be in reStructuredText markup.

e.g. https://pypi.python.org/pypi/ncbi-genome-download/0.2.0
vs https://pypi.python.org/pypi/ncbi-genome-download/0.2.1