core: Break the API so all list types now use the plural form.
Since the initial release, a lot of options were changed from accepting a
single parameter to accepting a list of parameters.
In order to not break the API and CLI, we decided to not update the option
names from singular to plural.

But as the upcoming version 0.3.0 is a bigger change from a programming
perspective anyway, we decided to update the option names to have them
more intuitively reflect the fact that they accept multiple parameters.

The CLI should still work, as ArgumentParser works on prefix matches, so
all plural forms that just append "s" still work, and compatibility
options for the two plurals that don't follow that pattern were added.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>